기본공격 구현완료 !

+++ Test/src/Monster.java
@@ -6,7 +6,7 @@ import javax.swing.ImageIcon;
 
 public class Monster {
 	int x,y; //ÁÂÇ¥
-	int hp; 
+	double hp; 
 	
 	Image monsterImg;
